refactor(saturation): relocate BacklogDriftConfig into sim/saturation (#1547 Part 2)
After Part 1 removed the dead batch analysis, BacklogDriftConfig had no consumer
inside sim/workload — its only users were in sim/saturation, which imported
upward for a type the owning package no longer used.

Move BacklogDriftConfig + NewBacklogDriftConfig + DefaultBacklogDriftConfig
verbatim into sim/saturation/backlog_drift_config.go (same 9 fields, same
validation panics, same defaults). Drop the workload. qualifier at every use
site in config.go and backlog_drift.go and remove the sim/workload import from
both. Move the 5 NewBacklogDriftConfig validation tests into the saturation
package; delete the now-empty sim/workload/saturation_test.go.

sim/saturation production files no longer import sim/workload at all (the only
surviving import is e2e_test.go's use of the unrelated TraceRecordsToRequestMetrics
trace helpers). sim/workload does not import sim/saturation (coupling severed,
not inverted). Config values unchanged, so stdout + saturation report are
byte-identical and --saturation-config YAML error messages are unchanged (INV-6).

Updates CLAUDE.md to reflect the removed batch API and the relocated config type.
