+++ sim/saturation/backlog_drift_config_test.go
@@ -1,5 +1,5 @@
-// sim/workload/saturation_test.go
-package workload
+// sim/saturation/backlog_drift_config_test.go
+package saturation
 
 import (
 	"fmt"
